chore: Update Next.js configuration for improved performance and security; rename app name in package.json; refactor layout and contact pages for better structure and SEO
